refactor: 타임딜 스케쥴러 관련 로직 분리
